Update Added documentation.
In ParseMetaMata() there need to be 2 yaml markers found in the document to try to identify the meta data. Previously the number just had to be greater than 0.

In YamlMarkers() now return a List<int> with 0 or 2 entries. 0 entries when zero or one marker was found. 2 entries if two markers were found.