Update: Implementing HTML builder/ Custom parser / GitMode change
A new HTML builder is now used to inject custom parsers and generate HTML from markdown. This builder is now the only place where Markdig is implemented.

Custom parsers, implementing IPipelineParser, can now be added to the generating process via the AddParser method. This method supports chaining.

GitMode has now a default value of false and is no longer initialized in the constructor.